Fix print formatting

+++ bench/packet_bench.cc
@@ -44,7 +44,7 @@ class rate_limiter {
     local_ops_ = 0;
     last_ts_ = high_resolution_clock::now();
     tspec_.tv_sec = 0;
-    fprintf(stderr, "10000 ops per %" PRIu64 " ns.\n", min_ns_per_10000_ops);
+    fprintf(stderr, "10000 ops per %lld ns.\n", min_ns_per_10000_ops);
   }
 
   uint64_t insert_packet(unsigned char* data, uint16_t len, token_list& tkns) {
@@ -71,7 +71,7 @@ class rate_limiter {
   struct timespec tspec_;
   high_resolution_clock::time_point last_ts_;
   uint64_t local_ops_;
-  uint64_t min_ns_per_10000_ops;
+  unsigned long long min_ns_per_10000_ops;
   packet_store::handle* handle_;
 };
